Move & Optimizing ping replies
